requirements.txt
openai>=1.51.0
openai-agents>=0.2.0

python-dotenv>=1.0.0
pydantic>=2.6.0
duckduckgo_search>=5.0
aiohttp>=3.9.0
requests>=2.31.0

gradio>=4.40.0